requirements.txt
airflow
pandas
SQLAlchemy==1.4.3
psycopg2-binary
psycopg2
python-dotenv
matplotlib
mysql-connector
line-profiler